Make it clear that tool references in package index result in installation
The platforms containing referenced cores, variants, and even tools in the case of manual platform installation, must have been installed. This could lead a platform author to think the same applies to the tools referenced from other packages by the platform's toolsDependencies field in package_index.json and that this results in their platform having a dependency on the other package's platform which uses the tool.